Make the project compile again with most code in place

+++ runtime_openxr/src/types.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 
 enum class GraphicsBackend {
     undefined = 0,
@@ -10,6 +11,6 @@ enum class GraphicsBackend {
 
 // Data types
 struct GBVector2i {
-    uint64_t x;
-    uint64_t y;
+    uint32_t x;
+    uint32_t y;
 };
